% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distribution.R
\name{check_distribution}
\alias{check_distribution}
\title{Simple check if distribution provided in defineUM() belongs
to a list of supported distributions.}
\usage{
check_distribution(object)
}
\arguments{
\item{object}{Any R object. In defineUM() it is used to
examine if selected distribution is in supported list of ditributions.}
}
\value{
TRUE or FALSE.
}
\description{
Simple check if distribution provided in defineUM() belongs
to a list of supported distributions.
}

