% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_Spatial.R
\name{check_if_Spatial}
\alias{check_if_Spatial}
\title{Simple check if class of provided object is Spatial}
\usage{
check_if_Spatial(object)
}
\arguments{
\item{object}{Any R object. In defineUM() it is used to
examine what type of data are delt with.}
}
\value{
TRUE or FALSE.
}
\description{
Simple check if class of provided object is Spatial
}
\author{
Kasia Sawicka
}
