% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defcat.R
\name{defcat}
\alias{defcat}
\title{Creating uncertainty model for categorical input variable}
\usage{
defcat(prob, . . .  )
}
\arguments{
\item{prob}{a data frame with listed probabilities for each category.}
}
\value{
\item{catprob}{ a data frame with list of probabilities per
category. }
}
\description{
Function that allows user to define uncertainty model of categorical
variable input.
}
\examples{

myexamples()

}
\author{
Kasia Sawicka, Gerard Heuvelink
}

