% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defjointnonspatial.R
\name{defjointnonspatial}
\alias{defjointnonspatial}
\title{Definition of joint non-spatial uncertainty model}
\usage{
defjointnonspatial(uncertain, means, varcov)
}
\arguments{
\item{uncert.objects}{Objects created by defnumarspatial}

\item{...}{}
}
\value{
Joint uncertainty model of cross-correlated spatial objects
}
\description{
Function to define joint uncertainty of cross-correlated spatial objects to do Monte Carlo.
}
\examples{

myjointnorm <- defjointnonspatial(uncertain = TRUE,
                                 means = c(5, 8, 10),
                                 varcov = matrix(c(1, 1.6, 0.2, 1.6, 4,
                                                   -2.4, 0.2, -2.4, 8),
                                                  nrow = 3, ncol = 3))
myjointnorm

}
\author{
Kasia Sawicka, Gerard Heuvelink
}

