% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defjointspatial.R
\name{defjointspatial}
\alias{defjointspatial}
\title{Definition of joint spatial uncertainty model}
\usage{
defjointspatial(uncert.objects, ...)
}
\arguments{
\item{uncert.objects}{Objects created by defnumarspatial}

\item{...}{}
}
\value{
Joint uncertainty model of cross-correlated spatial objects
}
\description{
Function to define joint uncertainty of cross-correlated spatial objects to do Monte Carlo.
}
\examples{

myjointnorm <- defjointnonspatial(uncertain = TRUE,
                                 means = c(5, 8, 10),
                                 varcov = matrix(c(1, 1.6, 0.2, 1.6, 4,
                                                   -2.4, 0.2, -2.4, 8),
                                                  nrow = 3, ncol = 3))
myjointnorm

}
\author{
Kasia Sawicka, Gerard Heuvelink
}

