% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem30m-data.R
\docType{data}
\name{dem30m}
\alias{dem30m}
\title{Digital Elevation Model of Zlatibor region in Serbia.}
\format{SpatialGridDataFrames}
\source{
The Zlatibor dataset was kindly provided by Prof. Branislav Bajat
 from the University of Belgrade, Serbia.
}
\usage{
data(dem30m)
}
\description{
A dataset containing the mean and sd of an example Digital Elevation Model.
}
\keyword{datasets}

