% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem30m_sd-data.R
\docType{data}
\name{dem30m_sd}
\alias{dem30m_sd}
\title{Standard deviation of Digital Elevation Model of Zlatibor region in Serbia.}
\format{a SpatialGridDataFrame with 15000 rows and 1 variable:
\describe{
\item{Elevation_sd}{Standard deviation of Digital Elevation Model, in meters}
}}
\source{
The Zlatibor dataset was kindly provided by Prof. Branislav Bajat
 from the University of Belgrade, Serbia.
}
\usage{
data(dem30m_sd)
}
\description{
A dataset containing the sd of an example Digital Elevation Model. 
It was calculated from dem30m using terrain funtion from raster package (opt = 'roughness').
}
