% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{depth}
\alias{depth}
\title{Title depth}
\usage{
depth(List)
}
\arguments{
\item{List}{an object of class 'list'}
}
\value{
an integer; level of list nesting
}
\description{
Title depth
}
\examples{

a <- list(1,2)
depth(a)

a <- list(list(1, 2), 3)
depth(a)

}

