% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_sampling.R
\name{distribution_sampling}
\alias{distribution_sampling}
\title{Sampling from a given distribution}
\usage{
distribution_sampling(n, distribution, parameters)
}
\arguments{
\item{n}{number of sampling runs}

\item{distribution}{A string describing selected distribution. The same as a part of
the string following the "r" in each random variate generation function in ?distributions.}

\item{parameters}{parameters to pass to the random variate generation function after
number of observations.}
}
\value{
Sample of random deviates.
}
\description{
Sampling from a given distribution
}
\author{
Kasia Sawicka
}

