% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executable.R
\name{executable}
\alias{executable}
\title{Wrapper function for calling executables in R}
\usage{
executable(filename)
}
\arguments{
\item{filename}{a path with a name to the .exe file to be wrapped here.}
}
\value{
Executable output.
}
\description{
Wrapper function for calling executables in R
}
\author{
Dennis Walvoort
}

