% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_strata.R
\name{find_strata}
\alias{find_strata}
\title{Sampling from a given distribution}
\usage{
find_strata(p, distribution, parameters, ...)
}
\arguments{
\item{p}{}

\item{distribution}{}

\item{parameters}{}
}
\value{
quantiles
}
\description{
Sampling from a given distribution
}

