% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.crm.R
\name{plot.crm}
\alias{plot.crm}
\title{Plots correlogram model.}
\usage{
\method{plot}{crm}(crm, distance = 1, ylim = c(0, 1), ...)
}
\arguments{
\item{crm}{Object of class "crm" as created by makecrm()}

\item{distance}{}

\item{ylim}{}

\item{...}{Additional parameters.}
}
\value{
plot of correlogram model
}
\description{
Plots correlogram model.
}
\examples{

mycormodel <- makecrm(acf0 = 0.8, range = 300, model = "Exp")
plot(mycormodel, distance = 1)
 
}

