% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotglyphs.R
\name{plotGlyphs}
\alias{plotGlyphs}
\title{Visualise Spatial Uncertainty using Glyphs.}
\usage{
plotGlyphs(simulations, density, title = NULL, xlab = NULL, ylab = NULL,
  grid = T, axis = T, mean_col = NULL, glyph_col = NULL,
  legend = NULL)
}
\description{
Allows for visualising spatial uncertainty for continuous data using glyphs.
}
\details{
@param simulations Object of class simulations containing continuous data
  @param density Numeric. Gives spacing of density of glyphs.
  @param title,xlab,ylab Character vector giving plot title, x axis label and y
    axis label respectively.
  @param grid Logical. Add grid to background of graph
  @param axis Logical. Add axis ticks and labels to graph
  @param mean_col \code{\link[grDevices]{colorRampPalette}}. Colours to
    be used for plotting the mean values of the realisations.
  @param glyph_col Character giving the colour of the glyphs representing the
    standard deviation of the realisations.
  @param legend Character giving the legend titles. To be provided in format
    c("legend1", "legend2").

@return
  @export

@examples
}

