% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.template.R
\name{print.template}
\alias{print.template}
\title{Print method for class "template."}
\usage{
\method{print}{template}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
what template the the template file contains.
}
\description{
Print method for class "template."
}
\examples{

}

