% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate1.R
\name{propagate1}
\alias{propagate1}
\title{Executing selected model runs with uncertain inputs}
\usage{
propagate(model, n, ...)
}
\arguments{
\item{realizations}{A list of samples. If contains more than one item, 
they have to be of the same number of realizations.}

\item{model}{}

\item{n}{number of model runs. Relates to number of first realizations to take.}

\item{...}{Additional parameters, e.g. model inputs/parameters. Must be named 
like in model definition.}
}
\value{
Multiple model runs as a list.
}
\description{
Function that allows user to run selected model number of times using MC
sample.
}
\details{
\strong{model} has to be a user specific function that uses input sample as
an argument.
}
\examples{



}
\author{
Kasia Sawicka, Dennis Walvoort, Gerard Heuvelink
}

