% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render}
\alias{render}
\title{Generaic function render.}
\usage{
render(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
cos
}
\description{
Rendering is the process of replacing the tags in moustaches by text.
For this, we provide a set of render-methods. See the `whisker`-package
(or https://mustache.github.io) for more information.
}
\examples{

}

