% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.template.R
\name{render.template}
\alias{render.template}
\title{Render method for "template" class.}
\usage{
\method{render}{template}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\value{
cos
}
\description{
Render method for "template" class.
}
\examples{

}

