% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleReshuffle.R
\name{sampleReshuffle}
\alias{sampleReshuffle}
\title{Sample reshuffle}
\usage{
sampleReshuffle(sample_matrix)
}
\arguments{
\item{sample_matrix}{}
}
\value{

}
\description{
Sample reshuffle
}
\examples{

t1 <- matrix(1:100, 20, 5)
sampleReshuffle(t1)

}

