% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.default.R
\name{sd.default}
\alias{sd.default}
\title{sd default}
\usage{
\method{sd}{default}(realizations, ...)
}
\arguments{
\item{realizations}{}

\item{...}{}
}
\value{
stats::sd
}
\description{
sd default
}
\examples{

}

