% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd}
\alias{sd}
\title{Standard Deviation}
\usage{
sd(realizations, ...)
}
\arguments{
\item{...}{additional parameters.}

\item{x}{as for stats::sd}
}
\value{
Standard Deviation
}
\description{
Standard Deviation
}

