% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spup.R
\docType{package}
\name{spup--pkg}
\alias{spup--pkg}
\alias{_PACKAGE}
\alias{spup--pkg-package}
\title{spup - Package for spatial uncertainty propagation}
\description{
Facilitates uncertainty propagation analysis
using Monte Carlo methods. In particular, provides functions
that allow to do uncertainty analysis with spatial variables/models.
}
\author{
Kasia Sawicka
}
