% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratsamp.R
\name{stratsamp}
\alias{stratsamp}
\title{Stratified sampling for spatial variables}
\usage{
stratsamp(n, distribution, parameters, p)
}
\arguments{
\item{n}{sample size per stratum}

\item{p}{vector of probs}

\item{mu}{mean of normal distribution to be sampled}

\item{sigma}{sd of normal distribution to be sampled}
}
\value{
Sample of spatial variable. Matrix with n rows and length(p)-1 columns.
}
\description{
Stratified sampling for spatial variables
}

