% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratsampSpatial.R
\name{stratsampSpatial}
\alias{stratsampSpatial}
\title{Stratified sampling for spatial variables}
\usage{
stratsampSpatial(mu, sigma, n, p)
}
\arguments{
\item{mu}{mean of normal distribution to be sampled}

\item{sigma}{sd of normal distribution to be sampled}

\item{n}{sample size per stratum}

\item{p}{vector of probs}
}
\value{
Sample of spatial variable. Matrix with n rows and length(p)-1 columns.
}
\description{
Stratified sampling for spatial variables
}

