% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Construct class "template".}
\usage{
template(filenames)
}
\arguments{
\item{filenames}{Name on the model input file.}
}
\value{
An obejct of a class "template".
}
\description{
Class that stores all templates with model inputs. The aim of this class is to: 1. organise model input files;
2. perform some checks.
}
\details{
A template is simply a model input file with: 1. the additional extension `.template`. 
2. input that needs to be modified is replaced by mustache-style tags.
}
\examples{

my_template <- template("vignettes/examples/input.txt.template")
readLines(my_template, warn = FALSE)

}

