% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TN-data.R
\docType{data}
\name{TN}
\alias{TN}
\title{Soil total nitrogen content in a south area (33 x 33km) of lake Alaotra in Madagascar.}
\format{a RasterLayer with dimensions  : 134, 135, 18090  (nrow, ncol, ncell), 
resolution  : 250, 250  (x, y).}
\source{
ISRIC soilgrid information. HENGL, T., MENDES DE JESUS,
 J., HEUVELINK, G. B. M., RUIPEREZ GONZALEZ, M., KILIBARDA, M.,
  BLAGOTIC, A., SHANGGUAN, W., WRIGHT, M. N., GENG, X., BAUER-MARSCHALLINGER, B.,
   GUEVARA, M. A., VARGAS, R., MACMILLAN, R. A., BATJES, N. H., LEENAARS, J. G. B.,
    RIBEIRO, E., WHEELER, I., MANTEL, S. & KEMPEN, B. 2017. SoilGrids250m:
     Global gridded soil information based on machine learning. PLOS ONE, 12, e0169748.
}
\usage{
data(TN)
}
\description{
A dataset containing the mean of soil TN content from 0-30 cm layer.
}
\keyword{datasets}

