% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugs.R
\name{ugs}
\alias{ugs}
\title{ugs}
\usage{
ugs(n, mean, sd, mask, crm)
}
\arguments{
\item{n}{Number of draws}

\item{mean}{Map of mean values}

\item{sd}{Map of sd values}

\item{mask}{Empty map for creating the dummy map of residuals}

\item{crm}{Correlogram model. Output of makecrm()}
}
\description{
ugs
}
\examples{

}

