% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.default.R
\name{var.default}
\alias{var.default}
\title{Variance as in stats::var}
\usage{
\method{var}{default}(realizations, ...)
}
\arguments{
\item{realizations}{a numeric vector, matrix or data frame.}

\item{...}{additional parameters.}
}
\value{
Variance
}
\description{
Variance as in stats::var
}

