% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_MC_sgdf.R
\name{var_MC_sgdf}
\alias{var_MC_sgdf}
\title{var() function for MC sample saved in a SpatialGridDataFrame}
\usage{
var_MC_sgdf(realizations, ...)
}
\arguments{
\item{realizations}{MC sample saved in SpatialGridDataFrame.}

\item{...}{additional parameters.}
}
\value{
SpatialGridDataFrame; a variance of a MC sample.
}
\description{
Calculates var from MC realizations for each location in a map.
}
\examples{

data(dem30m, dem30m_sd)
dem_crm <- makecrm(acf0 = 0.78, range = 321, model = "Exp")
demUM <- defineUM(uncertain = TRUE, distribution = "norm", 
                   distr_param = c(dem30m, dem30m_sd), crm = dem_crm)
dem_sample <- genSample(UMobject = demUM, n = 50, samplemethod = "ugs",
                        nmax = 20, asList = FALSE)
dem_var <- var_MC_sgdf(dem_sample)

}
\author{
Kasia Sawicka
}

