% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcov.R
\name{varcov}
\alias{varcov}
\title{Calculate variance covariance matrix}
\usage{
varcov(sd_vector, cormat)
}
\arguments{
\item{sd_vector}{vector of standard deviations.}

\item{cormat}{correlation matrix.}
}
\value{
Variance-covariance matrix.
}
\description{
Calculate variance covariance matrix
}
\examples{
vc <- varcov(c(1,2,3), matrix(c(1,0.7,0.2,0.7,1,0.5,0.2,0.5,1), nrow = 3, ncol = 3))
vc

}
\author{
Kasia Sawicka
}

