% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgm2crm.R
\name{vgm2crm}
\alias{vgm2crm}
\title{Convert vgm to crm}
\usage{
vgm2crm(vgm, psill, nugget, range, model, kappa = 0.5, anis, Err = 0)
}
\arguments{
\item{vgm}{See ?vgm}

\item{psill}{See ?vgm}

\item{nugget}{See ?vgm}

\item{range}{See ?vgm}

\item{model}{See ?vgm}

\item{kappa}{See ?vgm}

\item{anis}{See ?vgm}

\item{Err}{See ?vgm}
}
\value{
Spatial correlogram model - standardised parameters of spatial variogram model
}
\description{
Convert vgm to crm
}
\author{
Kasia Sawicka
}
