% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woon-data.R
\docType{data}
\name{woon}
\alias{woon}
\title{Neighbourhood in Rotterdam.}
\format{a SpatialPolygonDataFrame with 723 polygons and 7 variables:
\describe{
\item{vbos}{number of addresses present in the building}
\item{woonareash}{residential area, in percent}
\item{Function}{assigned category depending on vbos and woonareash -
 for residential is 1, for offcie is 2, for other is 3}
\item{residential}{probability that the building is residential}
\item{office}{probability that the building is an office}
\item{other}{probability that the building has other function}
\item{check}{check if probabilities sum to 1}
}}
\source{
Kadaster, NL.
}
\usage{
data(dem30m_sd)
}
\description{
The 'woon' object is a SpatialPolygonDataFrame where each building is represented by one polygon.
}
\keyword{datasets}

