\name{defcat}
\alias{defcat}
\title{Creating uncertainty model for categorical input variable}

% Description section
\description{
Function that allows user to define uncertainty model of categorical variable input.
}

% Usage section
\usage{
defcat(prob, . . .  )
}

% Arguments section
\arguments{
  \item{prob}{ a data frame with listed probabilities for each category. }
}

% Details section
\details{}

% Value section
\value{
  \item{catprob}{ a data frame with list of probabilities per category. }
}

% Author(s) section
\author{ Kasia Sawicka, Gerard Heuvelink}

% References section
\references{}

% Examples section
\examples{
myexamples()
}