% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMUM.r
\name{defineMUM}
\alias{defineMUM}
\title{Define Mulivariate Uncertainty Model}
\usage{
defineMUM(UMlist, cormatrix, ...)
}
\arguments{
\item{UMlist}{a list of uncertain objects creaded in defineUM()}

\item{cormatrix}{matrix of cross-correlations}

\item{...}{}
}
\value{
Object of a class "JointNumericSpatial" or
}
\description{
In case of scalar a square matrix of correlations,
dimentionally equal to the number of objects, square, symetric 
(transposed must be the same as original, diagonal must all be 1
all values must be <-1, +1>) and all eigenvalues must be > 0.
}
\examples{

data(Madagascar)
OC_crm <- makecrm(acf0 = 0.6, range = 1000, model = "Sph")
OC_UM <- defineUM(TRUE, distribution = "norm", 
                  distr_param = c(OC, OC_sd), crm = OC_crm,
                  id = "OC", cross_ids = "TN")
class(OC_UM)

TN_crm <- makecrm(acf0 = 0.4, range = 1000, model = "Sph")
TN_UM <- defineUM(TRUE, distribution = "norm", 
                  distr_param = c(TN, TN_sd), crm = TN_crm,
                  id = "TN", cross_ids = "OC")
class(TN_UM)

OC <- list(OC_UM,TN_UM)
class(OC)
length(OC)
str(OC)
 
test1 <- defineMUM(OC,matrix(c(1,0.7,0.7,1), nrow=2, ncol=2))
test2 <- defineMUM(OC,matrix(c(1,0.7,0.6,1), nrow=2, ncol=2))
test3 <- defineMUM(OC,matrix(c(1,-0.7,-0.7,1.1), nrow=2, ncol=2))                  
test3 <- defineMUM(OC,matrix(c(1,1,1,1), nrow=2, ncol=2))                  

}

