% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineUMs.R
\name{defineUMs}
\alias{defineUMs}
\title{Creating an uncertainty model for a single input.}
\usage{
defnummarspatial(uncertain = TRUE, sp.obj = NULL, crm = NULL, mask =
NULL, ...)
}
\arguments{
\item{uncertain}{"TRUE" or "FALSE", determines if specification of
Uncertainty Model (UM) is needed.}

\item{distribution}{A string specified which distribution to sample from.
See Details for a list of supported distributions.}

\item{distr_param}{A vector or a list with distribution parameters. For example, for 
normal distribution in spatial variable this must be a map of means and a map
of syandard deviations.}

\item{cat_prob}{A list of probabilities for categorical variables. In case of spatial inputs,
these must be maps. In case of non-spatial inputs it is a data.frame.}

\item{crm}{A correlogram model, object of a class "SpatialCorrelogramModel",
output of makecormodel(). Can only be specified for numerical variables.}

\item{...}{Additional parameters of um that may want to be saved here.}
}
\value{
A list of all necessary information for creating realizations of
the uncertain variable.
}
\description{
Function that allows user to define marginal uncertainty distributions 
for model inputs and subsequent Monte Carlo analysis.
}
\details{
\strong{uncertain} If "TRUE" the uncertainty model for the input has to be
specified. If uncertain ="FALSE" the function requires a mean value of a
distribution, e.g. a scalar, a vector, or a map.

The spatial object must contain a map of mean and standard deviation. If crm
is provided and spatial correlation between the residuals is assumed only
the normal distribution of residuals is allowed.

If no spatial correlations between residuals is assumed, allowed
distributions for marginal uncertainty models are listed in Table 1.



Table 1 Parametric probability models allowed in defineUM.
\tabular{rlllll}{ \tab \strong{Numeric type} \tab \strong{Distribution} \tab
\strong{Syntax} \tab \strong{Parameters} \tab \strong{Description} \cr \tab
Continuous \tab Beta \tab "beta" \tab c(\eqn{\alpha}, \eqn{\beta}) \tab
Shape parameters \eqn{\alpha} > 0, \eqn{\beta} > 0 \cr \tab Continuous \tab
Cauchy \tab \tab \tab \cr \tab Continuous \tab Chi-squared \tab \tab \tab
\cr \tab Continuous \tab Continuous uniform \tab \tab \tab \cr \tab
Continuous \tab Exponential \tab \tab \tab \cr \tab Continuous \tab Gamma
\tab \tab \tab \cr \tab Continuous \tab \tab \tab \tab

}
}
\examples{

}
\author{
Kasia Sawicka, Gerard Heuvelink
}

