\name{defjointspatial}
\alias{defjointspatial}
\title{Definition of joint spatial uncertainty model}

% Description section
\description{

}

% Usage section
\usage{

}

%Arguments section
\arguments{

}

% Details section
\details{

}

% Value section
\value{

}

% Author(s) section
\author{ Kasia Sawicka, Gerard Heuvelink}

% References section
\references{}

% Examples section
\examples{
myexamples()
}
