% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem30m-data.R
\docType{data}
\name{dem30m}
\alias{dem30m}
\title{Digital Elevation Model of Zlatibor region in Serbia.}
\format{a SpatialGridDataFrame with 15000 rows and 1 variable:
\describe{
\item{Elevation}{Digital Elevation Model, in meters}
}}
\source{
The Zlatibor dataset was kindly provided by Prof. Branislav Bajat
 from the University of Belgrade, Serbia.
}
\usage{
data(dem30m)
}
\description{
A dataset containing the mean an example Digital Elevation Model.
}
\keyword{datasets}

