% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_sampling_raster.R
\name{distribution_sampling_raster}
\alias{distribution_sampling_raster}
\title{Sampling from a given distribution}
\usage{
distribution_sampling_raster(distribution, parameters_stack)
}
\arguments{
\item{distribution}{A string describing selected distribution. The same as a part of
the string following the "r" in each random variate generation function in ?distributions.}

\item{n}{number of observations}

\item{parameters}{parameters to pass to the random variate generation function after
number of observations.}
}
\value{
sample of random deviates
}
\description{
Sampling from a given distribution
}

