% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_strata.R
\name{find_strata}
\alias{find_strata}
\title{Sampling from a given distribution}
\usage{
find_strata(p, distribution, parameters, ...)
}
\arguments{
\item{p}{a vector of quantiles.}

\item{distribution}{a string indicating which distribution to sample from.
See ?defineUM() for Details.}

\item{parameters}{parameters to pass to the appropriate sampling funtion, e.g.
mean and sd for "norm" distribution.}

\item{...}{additional parameters.}
}
\value{
Strata of the distribution defined by given quantiles.
}
\description{
Sampling from a given distribution
}
\author{
Kasia Sawicka, Stefan van Dam
}

