% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSample.MarginalCategoricalSpatial.R
\name{genSample.MarginalCategoricalSpatial}
\alias{genSample.MarginalCategoricalSpatial}
\title{Generating Monte Carlo sample from an uncertain object of a class 
'MarginalCategoricalSpatial'}
\usage{
genSample(UMobject, n, ...)
}
\arguments{
\item{UMobject}{uncertain object defined using defineUM().}

\item{n}{Integer. Number of Monte Carlo realizations.}

\item{...}{additional parameters}
}
\value{
A Monte Carlo sample of a categorical spatial variable.
}
\description{
Generating Monte Carlo sample from an uncertain object of a class 
'MarginalCategoricalSpatial'
}
\examples{

# load data
data(house)
houseUM <- defineUM(uncertain = TRUE, categories = c(100,200), cat_prob = houses_DF)
h_sample <- genSample(houseUM, n = 10)
str(h_sample)

houseUM <- defineUM(uncertain = TRUE, categories = c("yes", 1), cat_prob = houses_DF)

}
\author{
Kasia Sawicka
}

