% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSample.MarginalCategoricalSpatial.R
\name{genSample.MarginalCategoricalSpatial}
\alias{genSample.MarginalCategoricalSpatial}
\title{Generating Monte Carlo sample from an uncertain object of a class 
'MarginalCategoricalSpatial'}
\usage{
genSample(UMobject, n, ...)
}
\arguments{
\item{UMobject}{uncertain object defined using defineUM().}

\item{n}{Integer. Number of Monte Carlo realizations.}

\item{...}{additional parameters}
}
\value{
A Monte Carlo sample of a categorical spatial variable.
}
\description{
Generating Monte Carlo sample from an uncertain object of a class 
'MarginalCategoricalSpatial'
}
\examples{

# load data
data(house)
houseUM <- defineUM(uncertain = TRUE, categories = c(100,200), cat_prob = houses_DF)
h_sample <- genSample(houseUM, n = 10)
str(h_sample)

data(Rotterdam)
woonUM <- defineUM(TRUE, categories = c(1,2,3), cat_prob = woon[, c(4:6)])
# woon_sample <- genSample(woonUM, 10, asList=FALSE)
# class(woon_sample)
# str(woon_sample@data)
# spplot(woon_sample)
woon_sample <- genSample(woonUM, 10)
class(woon_sample)

}
\author{
Kasia Sawicka
}

