% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSample.R
\name{genSample}
\alias{genSample}
\title{Generic function for sampling.}
\usage{
genSample(UMobject, n, samplemethod, p = 0, ...)
}
\arguments{
\item{UMobject}{}

\item{n}{}

\item{samplemethod}{}

\item{p}{}

\item{...}{}
}
\value{
Monte Carlo sample
}
\description{
Generic function for sampling.
}
\examples{


}

