% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSample.R
\name{genSample}
\alias{genSample}
\title{Generic function for sampling.}
\usage{
genSample(UMobject, n, samplemethod, p = 0, asList = TRUE, ...)
}
\arguments{
\item{UMobject}{uncertain object}

\item{n}{number of Monte Carlo simulations}

\item{samplemethod}{sampling method}

\item{p}{number of strata}

\item{asList}{logical. If asList = TRUE returns list of all samples as a list. 
If asList = FALSE returns samples in a format of distribution parameters in UMobject.}

\item{...}{other parameters}
}
\value{
Monte Carlo sample
}
\description{
Generic function for sampling.
}
\examples{
# examples here

}

