% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.template.R
\name{print.template}
\alias{print.template}
\title{Print method for class "template."}
\usage{
print.template(x, ...)
}
\arguments{
\item{x}{Object of class "template".}

\item{...}{additional parameters.}
}
\value{
Template file content.
}
\description{
Print method for class "template."
}
\author{
Dennis Walvoort
}
