% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate.R
\name{propagate}
\alias{propagate}
\title{Title Propagation function}
\usage{
propagate(realizations, model, n, ...)
}
\arguments{
\item{realizations}{a list or a list of lists; max one nesting is allowed.}

\item{model}{a model that is written as a function in R}

\item{n}{number of Monte Carlo Runs}

\item{...}{any arguments that the model takes on top of realizations}
}
\value{
model output realizations
}
\description{
Title Propagation function
}
\examples{

data(DEM)
dem_crm <- makecrm(acf0 = 0.78, range = 321, model = "Exp")
demUM <- defineUM(uncertain = TRUE, distribution = "norm", 
                   distr_param = c(dem30m, dem30m_sd), crm = dem_crm)
dem_sample <- genSample(UMobject = demUM, n = 5, samplemethod = "ugs", nmax = 20)
slope_sample <- propagate(dem_sample, model = Slope, n = 5, projection = CRS("+init=epsg:3857"))

}

