% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate1.R
\name{propagate1}
\alias{propagate1}
\title{Executing selected model runs with uncertain inputs}
\usage{
propagate(model, n, ...)
}
\arguments{
\item{realizations}{A list of samples. If contains more than one item, 
they have to be of the same number of realizations.}

\item{model}{}

\item{n}{number of model runs. Relates to number of first realizations to take.}

\item{...}{Additional parameters, e.g. model inputs/parameters. Must be named 
like in model definition.}
}
\value{
Multiple model runs as a list.
}
\description{
Function that allows user to run selected model number of times using MC
sample.
}
\details{
\strong{model} has to be a user specific function that uses input sample as
an argument.
}
\examples{

data(DEM)
dem_crm <- makecrm(acf0 = 0.78, range = 321, model = "Exp")
demUM <- defineUMs(uncertain = TRUE, distribution = "norm", 
                   distr_param = c(dem30m, dem30m_sd), crm = dem_crm)
dem_sample <- genSample(UMobject = demUM, n = 50, samplemethod = "ugs", nmax = 20)
slope_sample <- propagate1(realizations = dem_sample, model = Slope, n = 50)

}
\author{
Kasia Sawicka, Dennis Walvoort, Gerard Heuvelink
}

