% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.SpatialGridDataFrame.R
\name{quantile.SpatialGridDataFrame}
\alias{quantile.SpatialGridDataFrame}
\title{quantile function for MC sample saved in SpatialGridDataFrame}
\usage{
\method{quantile}{SpatialGridDataFrame}(realizations, ...)
}
\arguments{
\item{realizations}{}

\item{...}{}
}
\value{
quantile of MC sample
}
\description{
quantile function for MC sample saved in SpatialGridDataFrame
}
\examples{

data(DEM)
dem_crm <- makecrm(acf0 = 0.78, range = 321, model = "Exp")
demUM <- defineUMs(uncertain = TRUE, distribution = "norm", 
                   distr_param = c(dem30m, dem30m_sd), crm = dem_crm)
dem_sample <- genSample(UMobject = demUM, n = 50, samplemethod = "ugs", nmax = 20)
slope_sample <- propagate1(realizations = dem_sample, model = Slope, n = 50)
slope_q <- quantile(slope_sample, probs = c(0.1, 0.9), na.rm = TRUE)


}

