% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.template.R
\name{render.template}
\alias{render.template}
\title{Render method for "template" class.}
\usage{
\method{render}{template}(x, ...)
}
\arguments{
\item{x}{an object of class "template", a model input file with additional
extension ".template".}

\item{...}{additional parameters.}
}
\value{
Rendered template file.
}
\description{
Rendering is the process of replacing the tags in moustaches by text.
}
\examples{

# examples here

}
\author{
Dennis Walvoort
}

