% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.default.R
\name{sd.default}
\alias{sd.default}
\title{Standard Deviation as in stats::sd}
\usage{
\method{sd}{default}(realizations, ...)
}
\arguments{
\item{realizations}{a numeric vector or an R object which is coercible to one by as.double(x).}

\item{...}{additional parameters.}
}
\value{
Standard Deviation
}
\description{
Standard Deviation as in stats::sd
}

