% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.SpatialGridDataFrame.R
\name{sd.SpatialGridDataFrame}
\alias{sd.SpatialGridDataFrame}
\title{Calculate standard deviation of the MC sample that is saved in a SpatialGridDataFrame}
\usage{
\method{sd}{SpatialGridDataFrame}(realizations, ...)
}
\arguments{
\item{realizations}{MC sample}

\item{...}{additional parameters}
}
\value{
sd of MC sample
}
\description{
Calculate standard deviation of the MC sample that is saved in a SpatialGridDataFrame
}
\examples{

data(DEM)
dem_crm <- makecrm(acf0 = 0.78, range = 321, model = "Exp")
demUM <- defineUMs(uncertain = TRUE, distribution = "norm", 
                   distr_param = c(dem30m, dem30m_sd), crm = dem_crm)
dem_sample <- genSample(UMobject = demUM, n = 50, samplemethod = "ugs", nmax = 20)
slope_sample <- propagate1(realizations = dem_sample, model = Slope, n = 50)
slope_sd <- sd(slope_sample, na.rm = TRUE)

}

