% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratsamp.R
\name{stratsamp}
\alias{stratsamp}
\title{Stratified sampling for spatial variables}
\usage{
stratsamp(n, distribution, parameters, p)
}
\arguments{
\item{n}{sample size per stratum.}

\item{distribution}{a string, distribution type to sample from.}

\item{parameters}{given distribution parameters.}

\item{p}{a vector of quantiles.}
}
\value{
Sample of spatial variable. Matrix with n rows and length(p)-1 columns.
}
\description{
Stratified sampling for spatial variables
}
\author{
Stefan van Dam, Kasia Sawicka
}

