% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.default.R
\name{var.default}
\alias{var.default}
\title{var default}
\usage{
\method{var}{default}(realizations, ...)
}
\arguments{
\item{realizations}{}

\item{...}{}
}
\value{
stats::var
}
\description{
var default
}
\examples{

}

